ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pandas==2.1.1
scikit-learn==1.3.1
psycopg2-binary==2.9.9
redis==5.0.1
kafka-python==2.0.2
python-dotenv==1.0.0
joblib==1.3.2
pytest==7.4.2
httpx==0.25.0